fixing tags

+++ create-license.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
 #
+
 cfos_license_input_file="CFOSVLTM24000026.lic"
-[[ -f $cfos_license_input_file ]] || echo $cfos_license_input_file does not exist
+
+if [ ! -f "$cfos_license_input_file" ]; then
+    echo "${cfos_license_input_file} does not exist"
+    exit 1
+fi
+
 file="manifests/base/store-firewall-license.yaml"
 licensestring=$(sed '1d;$d' $cfos_license_input_file | tr -d '\n')
 cat <<EOF >$file
